requirements.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==2.3.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-schema-adapter==3.0.6
drf-spectacular==0.27.2
environ==1.0
inflection==0.5.1
# Inflector==3.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1